Time: 0 ms (100.00%) | Memory: 7.9 MB (34.82%) - LeetSync

+++ 2050-count-good-numbers/count-good-numbers.cpp
@@ -2,22 +2,26 @@ class Solution {
 public:
     const int MOD = 1e9+7;
     long long calculatePower(long long a, long long b){
+        // Base case
         if(b == 0) return 1;
+        
         long long half = calculatePower(a, b/2);
-        long long result = (half*half) % MOD;
-        if(b % 2 == 1){
-            result = (result*a)%MOD;
+        long long result = (half*half)%MOD;
+
+        if(b%2 == 1){
+            result = (result*a) % MOD;
         }
         return result;
     }
-    int countGoodNumbers(long long n) {
+    long long countGoodNumbers(long long n) {
         long long evenCount = (n+1)/2;
         long long oddCount = n/2;
-
-        long long result = (calculatePower(5, evenCount)* calculatePower(4, oddCount)) % MOD;
+        long long result = (calculatePower(5, evenCount)*calculatePower(4, oddCount)) % MOD;
         return result;
     }
 };
+
+
 
 
 
@@ -32,7 +36,7 @@ public:
 //                 return false;
 //             }
 //         }
-//         return false;
+//         return true;
 //     }
 //     int countGoodNumbers(long long n) {
 //         long long start = pow(10, n-1);
